// payload-types.ts
export interface TeamMember {
  id: string;
  name: string;
  roles: (string | TeamRole)[];
  avatar: string;
  github?: string | null;
  linkedin?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "team-roles".
 */
export interface TeamRole {
  id: string;
  name: string;
  slug?: string | null;
  priority?: number | null;
  description?: string | null;
  updatedAt: string;
  createdAt: string;
}